# WMS.Angular/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WMS API v1",
    "version": "v1"
  },
  "paths": {
...
    "/manualpicking": {
      "get": {
        "tags": [
          "WMS.Api"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManualPickingSummaryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "WMS.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManualPickingDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "WMS.Api"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "WMS.Api"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/manualpicking/{Id}": {
      "get": {
        "tags": [
          "WMS.Api"
        ],
        "operationId": "GetPick",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateManualPickingDto": {
        "type": "object",
        "properties": {
          "warehouseId": {
            "type": "integer",
            "format": "int32"
          },
          "binId": {
            "type": "integer",
            "format": "int32"
          },
          "checkInId": {
            "type": "integer",
            "format": "int32"
          },
          "pickedProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PickedProductDetailsDto"
            },
            "nullable": true
          },
          "pickingDate": {
            "type": "string",
            "format": "date-time"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "ManualPickingSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "warehouse": {
            "type": "string",
            "nullable": true
          },
          "bin": {
            "type": "string",
            "nullable": true
          },
          "pickProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PickedProductSummaryDto"
            },
            "nullable": true
          },
          "pickingDate": {
            "type": "string",
            "format": "date-time"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PickedProductDetailsDto": {
        "type": "object",
        "properties": {
          "receivedProductId": {
            "type": "integer",
            "format": "int32"
          },
          "quantityPicked": {
            "type": "integer",
            "format": "int32"
          },
          "datePicked": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "PickedProductSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "quantityPicked": {
            "type": "integer",
            "format": "int32"
          },
          "datePicked": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "CookieAuth": {
        "type": "apiKey",
        "description": "ASP.NET Core Identity Application Cookie",
        "name": ".Skybest.WMS.Auth",
        "in": "cookie"
      }
    }
  },
  "security": [
    {
      "CookieAuth": []
    }
  ]
}